+++ app/tests/PhpTest.php
@@ -5,12 +5,9 @@ namespace Tests;
 use Helldar\PrettyArray\Services\File as Pretty;
 use Helldar\Support\Facades\Helpers\Filesystem\Directory;
 use Helldar\Support\Facades\Helpers\Filesystem\File;
-use Helldar\Support\Facades\Helpers\Str;
 
 final class PhpTest extends TestCase
 {
-    protected string $target_path = __DIR__ . '/../../src';
-
     public function testPhp(): void
     {
         foreach ($this->files() as $filename) {
@@ -24,7 +21,7 @@ final class PhpTest extends TestCase
                 $this->assertSame(array_keys($source), array_keys($target), $this->messageForLocale($path, $locale));
 
                 if ($this->isInline($filename)) {
-                    $this->assertDoesntSee($path, ':attribute');
+                    $this->assertDoesntSee($path, [':attribute', ':Attribute']);
                 }
             }
         }
@@ -32,7 +29,7 @@ final class PhpTest extends TestCase
 
     protected function files(): array
     {
-        return File::names($this->source_path, static fn ($filename) => Str::endsWith($filename, '.php'));
+        return File::names($this->source_path, static fn ($filename) => str_ends_with($filename, '.php'));
     }
 
     protected function locales(): array