+++ app/tests/TestCase.php
@@ -4,7 +4,6 @@ namespace Tests;
 
 use Helldar\PrettyArray\Services\File;
 use Helldar\Support\Facades\Helpers\Arr;
-use Helldar\Support\Facades\Helpers\Str;
 use PHPUnit\Framework\TestCase as BaseTestCase;
 use Tests\Concerns\Messages;
 
@@ -12,9 +11,9 @@ abstract class TestCase extends BaseTestCase
 {
     use Messages;
 
-    protected string $source_path = __DIR__ . '/../../script/en';
+    protected string $source_path = __DIR__ . '/../../source';
 
-    protected string $target_path;
+    protected string $target_path = __DIR__ . '/../../locales';
 
     protected function source(string $filename): array
     {
@@ -39,19 +38,23 @@ abstract class TestCase extends BaseTestCase
         $file = File::make()->loadRaw($path);
 
         $this->assertTrue(
-            Str::contains($file, $content),
+            str_contains($file, $content),
             $this->messageAssertSee($path, $content)
         );
     }
 
-    protected function assertDoesntSee(string $path, string $content): void
+    protected function assertDoesntSee(string $path, string|array $content): void
     {
         $file = File::make()->loadRaw($path);
 
-        $this->assertFalse(
-            Str::contains($file, $content),
-            $this->messageAssertDoesntSee($path, $content)
-        );
+        $values = Arr::wrap($content);
+
+        foreach ($values as $value) {
+            $this->assertFalse(
+                str_contains($file, $value),
+                $this->messageAssertDoesntSee($path, $value)
+            );
+        }
     }
 
     protected function sort(array &$array): void
@@ -61,11 +64,11 @@ abstract class TestCase extends BaseTestCase
 
     protected function isValidation(string $filename): bool
     {
-        return Str::startsWith($filename, 'validation');
+        return str_starts_with($filename, 'validation');
     }
 
     protected function isInline(string $filename): bool
     {
-        return Str::contains($filename, 'inline');
+        return str_contains($filename, 'inline');
     }
 }